# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/customers": {
      "post": {
        "description": "Always creates a new customer (there is no lookup by email). The response masks the personal data.",
        "operationId": "customersCreate",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Client-generated UUID. Repeating it with the same body returns the stored response with `Idempotent-Replayed: true`; with another body, 409 IDEMPOTENCY_KEY_CONFLICT.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "headers": {
              "Location": {
                "description": "URL of the new customer.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters.\n\nInvalid request parameters.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with the current state or a previous request.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Create the customer of a checkout",
        "tags": [
          "Customers"
        ]
      }
    },
    "/api/v1/customers/{customerId}": {
      "get": {
        "operationId": "customersDetail",
        "parameters": [
          {
            "name": "customerId",
            "required": true,
            "in": "path",
            "description": "Customer id.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "The resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Get a customer with its personal data masked",
        "tags": [
          "Customers"
        ]
      }
    }
  },
  "info": {
    "title": "Tech Store Checkout API",
    "description": "Catalog, delivery coverage and checkout of a tech accessories store. Errors follow RFC 9457 Problem Details.",
    "version": "0.2.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Same origin (CloudFront routes /api/* to the API)"
    }
  ],
  "components": {
    "schemas": {
      "FieldError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "example": "productId"
          },
          "message": {
            "type": "string",
            "example": "Must be a valid UUID."
          }
        },
        "required": [
          "field",
          "message"
        ]
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "/problems/product-not-found"
          },
          "title": {
            "type": "string",
            "example": "Not Found"
          },
          "status": {
            "type": "number",
            "example": 404
          },
          "detail": {
            "type": "string",
            "example": "The product does not exist."
          },
          "instance": {
            "type": "string",
            "example": "/api/v1/products/7d094266-0b4e-4789-9522-96e1cd7ffa61"
          },
          "code": {
            "type": "string",
            "description": "Stable, machine-readable error code.",
            "example": "PRODUCT_NOT_FOUND"
          },
          "traceId": {
            "type": "string",
            "description": "Request id, also returned in the X-Request-Id header."
          },
          "context": {
            "type": "object",
            "description": "Extra data about the error.",
            "additionalProperties": {
              "oneOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                }
              ]
            }
          },
          "errors": {
            "description": "Invalid fields (400 only).",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "traceId"
        ]
      },
...
      "CreateCustomerRequest": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "description": "3 to 80 characters: letters (accents and ñ included), spaces, apostrophes, hyphens and dots; at least two words.",
            "example": "Ana María Gómez"
          },
          "email": {
            "type": "string",
            "description": "Up to 254 characters; stored in lower case.",
            "example": "ana.gomez@example.com"
          },
          "phone": {
            "type": "string",
            "description": "10 digits starting with 3 (mobile) or 60 (landline); +57, spaces and hyphens are accepted.",
            "example": "3001234567"
          },
          "legalIdType": {
            "type": "string",
            "enum": [
              "CC",
              "CE",
              "NIT",
              "PP"
            ],
            "example": "CC"
          },
          "legalId": {
            "type": "string",
            "description": "CC: 5-10 digits; CE: 6-10 letters or digits; NIT: 9 digits and an optional check digit; PP: 6-12 letters or digits.",
            "example": "1020304050"
          }
        },
        "required": [
          "fullName",
          "email",
          "phone",
          "legalIdType",
          "legalId"
        ]
      },
      "Customer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "2f9d4c1a-7b3e-4d5f-8a6b-9c0d1e2f3a4b"
          },
          "fullName": {
            "type": "string",
            "example": "Ana M. G."
          },
          "email": {
            "type": "string",
            "example": "a***@example.com"
          },
          "phone": {
            "type": "string",
            "example": "******4567"
          },
          "legalIdType": {
            "type": "string",
            "enum": [
              "CC",
              "CE",
              "NIT",
              "PP"
            ],
            "example": "CC"
          },
          "legalId": {
            "type": "string",
            "example": "******4050"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-24T20:15:00.000Z"
          }
        },
        "description": "Customer with its personal data masked.",
        "required": [
          "id",
          "fullName",
          "email",
          "phone",
          "legalIdType",
          "legalId",
          "createdAt"
        ]
      }
    }
  }
}